requirements.txt
altgraph==0.17.4
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
backoff==2.2.1
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==44.0.3
daphne==4.1.2
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
macholib==1.16.3
oauthlib==3.2.2
packaging==25.0
platformdirs==4.3.8
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.7.1
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
urllib3==2.4.0
zope.interface==7.2
